requirements.txt
Flask
openai
faiss-cpu
sentence-transformers
requests
langchain
langchain-openai
langchain-community
chromadb
python-dotenv
flask-cors
pypdf
PyPDF2